requirements.txt
numpy==1.20.0
umap-learn==0.5.0
numba==0.52
pandas==1.3.2
bertopic==0.9.4
tweet-preprocessor
sentence-transformers
nltk
